gst/vaapi: move AVC encoder test classes to lib
